requirements.txt
alembic==1.11.2
apispec==6.3.0
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.42
botocore==1.31.42
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-smorest==0.42.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==1.26.16
webargs==8.3.0
Werkzeug==2.3.6
zipp==3.16.2
